+++ CHANGELOG_PENDING.md
@@ -9,6 +9,17 @@
 * Apps
 
 * Go API
+- [mempool] \#2659 `Mempool` now an interface
+  * old `Mempool` implementation renamed to `CListMempool`
+  * `NewMempool` renamed to `NewCListMempool`
+  * `Option` renamed to `CListOption`
+  * unexpose `MempoolReactor.Mempool`
+  * `MempoolReactor` renamed to `Reactor`
+  * `NewMempoolReactor` renamed to `NewReactor`
+  * unexpose `TxID` method
+  * `TxInfo.PeerID` renamed to `SenderID`
+- [state] \#2659 `Mempool` interface moved to mempool package
+  * `MockMempool` moved to top-level mock package and renamed to `Mempool`
 - [libs/common] Removed `PanicSanity`, `PanicCrisis`, `PanicConsensus` and `PanicQ`
 - [node] Moved `GenesisDocProvider` and `DefaultGenesisDocProviderFunc` to state package
 
@@ -17,6 +28,7 @@
 * P2P Protocol
 
 ### FEATURES:
+- [node] \#2659 Add `node.Mempool()` method, which allows you to access mempool
 
 ### IMPROVEMENTS:
 - [rpc] [\#3534](https://github.com/tendermint/tendermint/pull/3534) Add support for batched requests/responses in JSON RPC
